// src/types/http.ts
export type AirplanesDotLive = {
  ac: {
    hex: string
    type: string
    flight?: string
    r?: string
    t?: string
    desc?: string
    alt_baro: any
    alt_geom?: number
    gs?: number
    ias?: number
    tas?: number
    mach?: number
    wd?: number
    ws?: number
    oat?: number
    tat?: number
    track?: number
    track_rate?: number
    roll?: number
    mag_heading?: number
    true_heading?: number
    baro_rate?: number
    geom_rate?: number
    squawk?: string
    emergency?: string
    category?: string
    nav_qnh?: number
    nav_altitude_mcp?: number
    nav_heading?: number
    nav_modes?: string[]
    lat: number
    lon: number
    nic: number
    rc: number
    seen_pos: number
    version?: number
    nic_baro?: number
    nac_p?: number
    nac_v?: number
    sil?: number
    sil_type?: string
    gva?: number
    sda?: number
    alert?: number
    spi?: number
    mlat: string[]
    tisb: any[]
    messages: number
    seen: number
    rssi: number
    dst: number
    dir: number
    ownOp?: string
    year?: string
    nav_altitude_fms?: number
    dbFlags?: number
    calc_track?: number
  }[]
  msg: string
  now: number
  total: number
  ctime: number
  ptime: number
}

export const enum DbFlags {
  Military = 1,
  Interesting = 2,
  PIA = 4,
  LADD = 8
}